HSEARCH-3125 Move IndexFamily to API
This requires to add a superinterface to IndexManagerType: IndexFamilyType.
That's because IndexManagerType is only SPI, and we need some type to use as
keys in the APIs allowing to retrieve an index family.
