+++ elasticsearch/src/main/java/org/hibernate/search/elasticsearch/indexes/ElasticsearchIndexFamily.java
@@ -4,9 +4,9 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.elasticsearch.indexes.spi;
+package org.hibernate.search.elasticsearch.indexes;
 
-import org.hibernate.search.indexes.spi.IndexFamily;
+import org.hibernate.search.indexes.IndexFamily;
 
 public interface ElasticsearchIndexFamily extends IndexFamily {
