+++ engine/src/main/java/org/hibernate/search/indexes/LuceneEmbeddedIndexFamilyType.java
@@ -1,0 +1,23 @@
+/*
+ * Hibernate Search, full-text search for your domain model
+ *
+ * License: GNU Lesser General Public License (LGPL), version 2.1 or later
+ * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
+ */
+package org.hibernate.search.indexes;
+
+import org.hibernate.search.indexes.spi.LuceneEmbeddedIndexManagerType;
+
+public final class LuceneEmbeddedIndexFamilyType {
+
+	private LuceneEmbeddedIndexFamilyType() {
+	}
+
+	/**
+	 * @return An {@link IndexFamilyType} representing the family of indexes using the embedded Lucene technology.
+	 */
+	public static IndexFamilyType get() {
+		return LuceneEmbeddedIndexManagerType.INSTANCE;
+	}
+
+}
