+++ engine/src/main/java/org/hibernate/search/indexes/impl/LuceneEmbeddedIndexFamilyImpl.java
@@ -17,8 +17,8 @@ import org.hibernate.search.engine.nulls.impl.LuceneMissingValueStrategy;
 import org.hibernate.search.engine.nulls.impl.MissingValueStrategy;
 import org.hibernate.search.engine.service.spi.ServiceManager;
 import org.hibernate.search.exception.SearchException;
+import org.hibernate.search.indexes.LuceneEmbeddedIndexFamily;
 import org.hibernate.search.indexes.spi.IndexFamilyImplementor;
-import org.hibernate.search.indexes.spi.LuceneEmbeddedIndexFamily;
 import org.hibernate.search.util.StringHelper;
 import org.hibernate.search.util.logging.impl.Log;
 import org.hibernate.search.util.logging.impl.LoggerFactory;
