+++ orm/src/main/java/org/hibernate/search/SearchFactory.java
@@ -9,7 +9,10 @@ package org.hibernate.search;
 import java.util.Set;
 
 import org.apache.lucene.analysis.Analyzer;
+
+import org.hibernate.search.indexes.IndexFamilyType;
 import org.hibernate.search.indexes.IndexReaderAccessor;
+import org.hibernate.search.indexes.IndexFamily;
 import org.hibernate.search.metadata.IndexedTypeDescriptor;
 import org.hibernate.search.query.dsl.QueryContextBuilder;
 import org.hibernate.search.stat.Statistics;
@@ -95,6 +98,13 @@ public interface SearchFactory {
 	 * @return the set of currently indexed types. If no types are indexed the empty set is returned.
 	 */
 	Set<Class<?>> getIndexedTypes();
+
+	/**
+	 * Get an {@link IndexFamily} using the index family type.
+	 * @param indexFamilyType the type of index family to retrieve
+	 * @return the selected {@link IndexFamily}, or null if it doesn't exist
+	 */
+	IndexFamily getIndexFamily(IndexFamilyType indexFamilyType);
 
 	/**
 	 * Unwraps some internal Hibernate Search types.
